requirements.txt
# Requirements for Product Image Automation Module

# Core dependencies
requests>=2.28.0
Pillow>=9.0.0
numpy>=1.21.0

# Optional dependencies for enhanced functionality
lxml>=4.6.0  # For XML parsing (Amazon API responses)
urllib3>=1.26.0  # For URL handling

# Development dependencies (optional)
pytest>=7.0.0  # For testing
responses>=0.18.0  # For mocking HTTP requests in tests